+++ finagle-core/src/test/java/com/twitter/finagle/StackParamCompilationTest.java
@@ -23,7 +23,6 @@ import com.twitter.finagle.loadbalancer.WhenNoNodesOpens;
 import com.twitter.finagle.naming.BindingFactory;
 import com.twitter.finagle.param.ExceptionStatsHandler;
 import com.twitter.finagle.param.Label;
-import com.twitter.finagle.param.Logger;
 import com.twitter.finagle.param.Monitor;
 import com.twitter.finagle.param.Reporter;
 import com.twitter.finagle.param.Stats;
@@ -56,7 +55,6 @@ public class StackParamCompilationTest {
       ClientBuilder.get()
         .configured(new Label("").mk())
         .configured(new Timer(DefaultTimer.getInstance()).mk())
-        .configured(new Logger(java.util.logging.Logger.getLogger("com.twitter.finagle")).mk())
         .configured(new Stats(com.twitter.finagle.stats.DefaultStatsReceiver.get()).mk())
         .configured(new Monitor(RootMonitor.getInstance()).mk())
         .configured(new Monitor(NullMonitor.getInstance()).mk())