+++ finagle-netty4-http/src/main/scala/com/twitter/finagle/netty4/http/package.scala
@@ -10,7 +10,7 @@ import com.twitter.finagle.netty4.http.handler.{
   UnpoolHttpHandler,
   UriValidatorHandler
 }
-import com.twitter.finagle.param.{Logger, Stats}
+import com.twitter.finagle.param.Stats
 import com.twitter.finagle.http.param._
 import com.twitter.finagle.server.Listener
 import com.twitter.finagle.transport.TransportContext
@@ -146,7 +146,6 @@ package object http {
     val maxRequestSize = params[MaxRequestSize].size
     val decompressionEnabled = params[Decompression].enabled
     val compressionLevel = params[CompressionLevel].level
-    val log = params[Logger].log
     val stats = params[Stats].statsReceiver
 
     { pipeline: ChannelPipeline =>